fix nav error and add comp banner
